// apps/server/src/modules/webhooks/enums/account-documents.enum.ts
export enum AccountDocumentsStatus {
  // Bank account approved.
  ACCOUNT_STATUS_BANK_ACCOUNT_INFO_APPROVED = "ACCOUNT_STATUS_BANK_ACCOUNT_INFO_APPROVED",

  // Bank account under review.
  ACCOUNT_STATUS_BANK_ACCOUNT_INFO_AWAITING_APPROVAL = "ACCOUNT_STATUS_BANK_ACCOUNT_INFO_AWAITING_APPROVAL",
  
  // Bank account status reverted to pending.
  ACCOUNT_STATUS_BANK_ACCOUNT_INFO_PENDING = "ACCOUNT_STATUS_BANK_ACCOUNT_INFO_PENDING",
  
  // Bank account rejected.
  ACCOUNT_STATUS_BANK_ACCOUNT_INFO_REJECTED = "ACCOUNT_STATUS_BANK_ACCOUNT_INFO_REJECTED",
  
  // Commercial information approved.
  ACCOUNT_STATUS_COMMERCIAL_INFO_APPROVED = "ACCOUNT_STATUS_COMMERCIAL_INFO_APPROVED",
  
  // Commercial information under review.
  ACCOUNT_STATUS_COMMERCIAL_INFO_AWAITING_APPROVAL = "ACCOUNT_STATUS_COMMERCIAL_INFO_AWAITING_APPROVAL",
  
  // Commercial information status reverted to pending.
  ACCOUNT_STATUS_COMMERCIAL_INFO_PENDING = "ACCOUNT_STATUS_COMMERCIAL_INFO_PENDING",
  
  // Commercial information rejected.
  ACCOUNT_STATUS_COMMERCIAL_INFO_REJECTED = "ACCOUNT_STATUS_COMMERCIAL_INFO_REJECTED",
  
  // Documents approved.
  ACCOUNT_STATUS_DOCUMENT_APPROVED = "ACCOUNT_STATUS_DOCUMENT_APPROVED",
  
  // Documents under review.
  ACCOUNT_STATUS_DOCUMENT_AWAITING_APPROVAL = "ACCOUNT_STATUS_DOCUMENT_AWAITING_APPROVAL",
  
  // Document status reverted to pending.
  ACCOUNT_STATUS_DOCUMENT_PENDING = "ACCOUNT_STATUS_DOCUMENT_PENDING",
  
  // Documents rejected.
  ACCOUNT_STATUS_DOCUMENT_REJECTED = "ACCOUNT_STATUS_DOCUMENT_REJECTED",
  
  // Account approved.
  ACCOUNT_STATUS_GENERAL_APPROVAL_APPROVED = "ACCOUNT_STATUS_GENERAL_APPROVAL_APPROVED",
  
  // Account under review.
  ACCOUNT_STATUS_GENERAL_APPROVAL_AWAITING_APPROVAL = "ACCOUNT_STATUS_GENERAL_APPROVAL_AWAITING_APPROVAL",
  
  // Account status reverted to pending.
  ACCOUNT_STATUS_GENERAL_APPROVAL_PENDING = "ACCOUNT_STATUS_GENERAL_APPROVAL_PENDING",
  
  // Account rejected.
  ACCOUNT_STATUS_GENERAL_APPROVAL_REJECTED = "ACCOUNT_STATUS_GENERAL_APPROVAL_REJECTED"
}